refactor: move graph-auspice conversion into separate modules
